refactor the common method and class
