+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/datafabric/dataset/service/mds/DatasetInstanceMDSUpgrader.java
@@ -24,7 +24,6 @@ import co.cask.cdap.api.dataset.lib.FileSet;
 import co.cask.cdap.api.dataset.lib.FileSetProperties;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
-import co.cask.cdap.common.io.Locations;
 import co.cask.cdap.data2.datafabric.dataset.DatasetMetaTableUtil;
 import co.cask.cdap.data2.datafabric.dataset.DatasetsUtil;
 import co.cask.cdap.data2.dataset2.DatasetFramework;
@@ -46,12 +45,11 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.List;
-import javax.annotation.Nullable;
 
 /**
  * Upgrades Dataset instances MDS
  */
-public final class DatasetInstanceMDSUpgrader {
+public final class DatasetInstanceMDSUpgrader extends AbstractMDSUpgrader {
   private static final Logger LOG = LoggerFactory.getLogger(DatasetInstanceMDSUpgrader.class);
   private final LocationFactory locationFactory;
   private final CConfiguration cConf;
@@ -91,6 +89,7 @@ public final class DatasetInstanceMDSUpgrader {
        });
   }
 
+  @Override
   public void upgrade() throws Exception {
     // Moves dataset instance meta entries into new table (in system namespace)
     // Also updates the spec's name ('cdap.user.foo' -> 'foo')
@@ -119,9 +118,8 @@ public final class DatasetInstanceMDSUpgrader {
    *
    * @param dsSpec the {@link DatasetSpecification} of the dataset
    * @return a boolean which is true if its a {@link FileSet} else false
-   * @throws IOException
    */
-  private boolean isFileSet(DatasetSpecification dsSpec) throws IOException {
+  private boolean isFileSet(DatasetSpecification dsSpec) {
     String dsType = dsSpec.getType();
     return (FileSet.class.getName().equals(dsType) || "fileSet".equals(dsType));
   }
@@ -185,33 +183,6 @@ public final class DatasetInstanceMDSUpgrader {
     } else {
       throw new IllegalArgumentException(String.format("Expected Dataset namespace to be either 'system' or 'user': %s",
                                                        dsId));
-    }
-  }
-  /**
-   * Renames the old location to new location if old location exists and the new one does not
-   *
-   * @param oldLocation the old {@link Location}
-   * @param newLocation the new {@link Location}
-   * @return new location if and only if the file or directory is successfully moved; null otherwise.
-   * @throws IOException
-   */
-  @Nullable
-  protected Location renameLocation(Location oldLocation, Location newLocation) throws IOException {
-    // if the newLocation does not exists or the oldLocation does we try to rename. If either one of them is false then
-    // the underlying call to renameTo will throw IOException which we propagate.
-    if (!newLocation.exists() || oldLocation.exists()) {
-      Locations.getParent(newLocation).mkdirs();
-      try {
-        return oldLocation.renameTo(newLocation);
-      } catch (IOException ioe) {
-        newLocation.delete();
-        LOG.warn("Failed to rename {} to {}", oldLocation, newLocation);
-        throw ioe;
-      }
-    } else {
-      LOG.debug("New location {} already exists and old location {} does not exists. The location might already be " +
-                  "updated.", newLocation, oldLocation);
-      return null;
     }
   }
 }
